requirements.txt
fastapi==0.68.3
uvicorn==0.15.0
torch==1.10.0
torchvision==0.11.1
facenet-pytorch==2.5.2
scikit-learn==0.24.2
Pillow==8.4.0
